processor/requirements.txt
pyspark==3.4.1
kafka-python==2.0.2
python-dotenv==1.0.0
cassandra-driver==3.25.0 